requirements.txt
pwntools
r2pipe